Categorize getters and setters into its own file
