+++ beacon-chain/state/stateV0/BUILD.bazel
@@ -6,14 +6,25 @@ go_library(
     srcs = [
         "cloners.go",
         "doc.go",
-        "field_root_attestation.go",
-        "field_root_eth1.go",
-        "field_root_validator.go",
-        "field_root_vector.go",
         "field_roots.go",
         "field_trie.go",
-        "getters.go",
-        "setters.go",
+        "field_trie_helpers.go",
+        "getters_attestation.go",
+        "getters_block.go",
+        "getters_checkpoint.go",
+        "getters_eth1.go",
+        "getters_misc.go",
+        "getters_randao.go",
+        "getters_state.go",
+        "getters_validator.go",
+        "setters_attestation.go",
+        "setters_block.go",
+        "setters_checkpoint.go",
+        "setters_eth1.go",
+        "setters_misc.go",
+        "setters_randao.go",
+        "setters_state.go",
+        "setters_validator.go",
         "state_trie.go",
         "types.go",
         "validator_getters.go",
@@ -61,7 +72,7 @@ go_test(
         "getters_test.go",
         "helpers_test.go",
         "references_test.go",
-        "setters_test.go",
+        "setters_attestation_test.go",
         "state_test.go",
         "state_trie_test.go",
         "types_test.go",
