--- beacon-chain/state/stateV0/field_root_eth1.go
@@ -1,59 +1,0 @@
-package stateV0
-
-import (
-	"github.com/pkg/errors"
-	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
-	"github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil"
-	"github.com/prysmaticlabs/prysm/shared/featureconfig"
-	"github.com/prysmaticlabs/prysm/shared/htrutils"
-)
-
-// eth1Root computes the HashTreeRoot Merkleization of
-// a BeaconBlockHeader struct according to the eth2
-// Simple Serialize specification.
-func eth1Root(hasher htrutils.HashFn, eth1Data *ethpb.Eth1Data) ([32]byte, error) {
-	if eth1Data == nil {
-		return [32]byte{}, errors.New("nil eth1 data")
-	}
-
-	enc := stateutil.Eth1DataEncKey(eth1Data)
-	if featureconfig.Get().EnableSSZCache {
-		if found, ok := cachedHasher.rootsCache.Get(string(enc)); ok && found != nil {
-			return found.([32]byte), nil
-		}
-	}
-
-	root, err := stateutil.Eth1DataRootWithHasher(hasher, eth1Data)
-	if err != nil {
-		return [32]byte{}, err
-	}
-
-	if featureconfig.Get().EnableSSZCache {
-		cachedHasher.rootsCache.Set(string(enc), root, 32)
-	}
-	return root, nil
-}
-
-// eth1DataVotesRoot computes the HashTreeRoot Merkleization of
-// a list of Eth1Data structs according to the eth2
-// Simple Serialize specification.
-func eth1DataVotesRoot(eth1DataVotes []*ethpb.Eth1Data) ([32]byte, error) {
-	hashKey, err := stateutil.Eth1DatasEncKey(eth1DataVotes)
-	if err != nil {
-		return [32]byte{}, err
-	}
-
-	if featureconfig.Get().EnableSSZCache {
-		if found, ok := cachedHasher.rootsCache.Get(string(hashKey[:])); ok && found != nil {
-			return found.([32]byte), nil
-		}
-	}
-	root, err := stateutil.Eth1DatasRoot(eth1DataVotes)
-	if err != nil {
-		return [32]byte{}, err
-	}
-	if featureconfig.Get().EnableSSZCache {
-		cachedHasher.rootsCache.Set(string(hashKey[:]), root, 32)
-	}
-	return root, nil
-}
