+++ beacon-chain/state/stateV0/field_trie.go
@@ -1,15 +1,11 @@
 package stateV0
 
 import (
-	"fmt"
 	"reflect"
 	"sync"
 
 	"github.com/pkg/errors"
-	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil"
-	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
-	"github.com/prysmaticlabs/prysm/shared/hashutil"
 )
 
 // FieldTrie is the representation of the representative
@@ -19,6 +15,7 @@ type FieldTrie struct {
 	reference   *stateutil.Reference
 	fieldLayers [][]*[32]byte
 	field       fieldIndex
+	length      uint64
 }
 
 // NewFieldTrie is the constructor for the field trie data structure. It creates the corresponding
@@ -30,6 +27,7 @@ func NewFieldTrie(field fieldIndex, elements interface{}, length uint64) (*Field
 			field:     field,
 			reference: stateutil.NewRef(1),
 			RWMutex:   new(sync.RWMutex),
+			length:    length,
 		}, nil
 	}
 	datType, ok := fieldMap[field]
@@ -40,6 +38,9 @@ func NewFieldTrie(field fieldIndex, elements interface{}, length uint64) (*Field
 	if err != nil {
 		return nil, err
 	}
+	if err := validateElements(field, elements, length); err != nil {
+		return nil, err
+	}
 	switch datType {
 	case basicArray:
 		return &FieldTrie{
@@ -47,6 +48,7 @@ func NewFieldTrie(field fieldIndex, elements interface{}, length uint64) (*Field
 			field:       field,
 			reference:   stateutil.NewRef(1),
 			RWMutex:     new(sync.RWMutex),
+			length:      length,
 		}, nil
 	case compositeArray:
 		return &FieldTrie{
@@ -54,6 +56,7 @@ func NewFieldTrie(field fieldIndex, elements interface{}, length uint64) (*Field
 			field:       field,
 			reference:   stateutil.NewRef(1),
 			RWMutex:     new(sync.RWMutex),
+			length:      length,
 		}, nil
 	default:
 		return nil, errors.Errorf("unrecognized data type in field map: %v", reflect.TypeOf(datType).Name())
@@ -77,6 +80,9 @@ func (f *FieldTrie) RecomputeTrie(indices []uint64, elements interface{}) ([32]b
 	}
 	fieldRoots, err := fieldConverters(f.field, indices, elements, false)
 	if err != nil {
+		return [32]byte{}, err
+	}
+	if err := f.validateIndices(indices); err != nil {
 		return [32]byte{}, err
 	}
 	switch datType {
@@ -106,6 +112,7 @@ func (f *FieldTrie) CopyTrie() *FieldTrie {
 			field:     f.field,
 			reference: stateutil.NewRef(1),
 			RWMutex:   new(sync.RWMutex),
+			length:    f.length,
 		}
 	}
 	dstFieldTrie := make([][]*[32]byte, len(f.fieldLayers))
@@ -118,6 +125,7 @@ func (f *FieldTrie) CopyTrie() *FieldTrie {
 		field:       f.field,
 		reference:   stateutil.NewRef(1),
 		RWMutex:     new(sync.RWMutex),
+		length:      f.length,
 	}
 }
 
@@ -138,114 +146,3 @@ func (f *FieldTrie) TrieRoot() ([32]byte, error) {
 	}
 }
 
-// this converts the corresponding field and the provided elements to the appropriate roots.
-func fieldConverters(field fieldIndex, indices []uint64, elements interface{}, convertAll bool) ([][32]byte, error) {
-	switch field {
-	case blockRoots, stateRoots, randaoMixes:
-		val, ok := elements.([][]byte)
-		if !ok {
-			return nil, errors.Errorf("Wanted type of %v but got %v",
-				reflect.TypeOf([][]byte{}).Name(), reflect.TypeOf(elements).Name())
-		}
-		return stateutil.HandleByteArrays(val, indices, convertAll)
-	case eth1DataVotes:
-		val, ok := elements.([]*ethpb.Eth1Data)
-		if !ok {
-			return nil, errors.Errorf("Wanted type of %v but got %v",
-				reflect.TypeOf([]*ethpb.Eth1Data{}).Name(), reflect.TypeOf(elements).Name())
-		}
-		return handleEth1DataSlice(val, indices, convertAll)
-	case validators:
-		val, ok := elements.([]*ethpb.Validator)
-		if !ok {
-			return nil, errors.Errorf("Wanted type of %v but got %v",
-				reflect.TypeOf([]*ethpb.Validator{}).Name(), reflect.TypeOf(elements).Name())
-		}
-		return stateutil.HandleValidatorSlice(val, indices, convertAll)
-	case previousEpochAttestations, currentEpochAttestations:
-		val, ok := elements.([]*pb.PendingAttestation)
-		if !ok {
-			return nil, errors.Errorf("Wanted type of %v but got %v",
-				reflect.TypeOf([]*pb.PendingAttestation{}).Name(), reflect.TypeOf(elements).Name())
-		}
-		return handlePendingAttestation(val, indices, convertAll)
-	default:
-		return [][32]byte{}, errors.Errorf("got unsupported type of %v", reflect.TypeOf(elements).Name())
-	}
-}
-
-func handleEth1DataSlice(val []*ethpb.Eth1Data, indices []uint64, convertAll bool) ([][32]byte, error) {
-	length := len(indices)
-	if convertAll {
-		length = len(val)
-	}
-	roots := make([][32]byte, 0, length)
-	hasher := hashutil.CustomSHA256Hasher()
-	rootCreator := func(input *ethpb.Eth1Data) error {
-		newRoot, err := eth1Root(hasher, input)
-		if err != nil {
-			return err
-		}
-		roots = append(roots, newRoot)
-		return nil
-	}
-	if convertAll {
-		for i := range val {
-			err := rootCreator(val[i])
-			if err != nil {
-				return nil, err
-			}
-		}
-		return roots, nil
-	}
-	if len(val) > 0 {
-		for _, idx := range indices {
-			if idx > uint64(len(val))-1 {
-				return nil, fmt.Errorf("index %d greater than number of items in eth1 data slice %d", idx, len(val))
-			}
-			err := rootCreator(val[idx])
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return roots, nil
-}
-
-func handlePendingAttestation(val []*pb.PendingAttestation, indices []uint64, convertAll bool) ([][32]byte, error) {
-	length := len(indices)
-	if convertAll {
-		length = len(val)
-	}
-	roots := make([][32]byte, 0, length)
-	hasher := hashutil.CustomSHA256Hasher()
-	rootCreator := func(input *pb.PendingAttestation) error {
-		newRoot, err := stateutil.PendingAttRootWithHasher(hasher, input)
-		if err != nil {
-			return err
-		}
-		roots = append(roots, newRoot)
-		return nil
-	}
-	if convertAll {
-		for i := range val {
-			err := rootCreator(val[i])
-			if err != nil {
-				return nil, err
-			}
-		}
-		return roots, nil
-	}
-	if len(val) > 0 {
-		for _, idx := range indices {
-			if idx > uint64(len(val))-1 {
-				return nil, fmt.Errorf("index %d greater than number of pending attestations %d", idx, len(val))
-			}
-			err := rootCreator(val[idx])
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return roots, nil
-}
